refactor(uds): consolidate UDS constants and NRC helpers into cda-interfaces
- Move service_ids, subfunction_ids, nrc, UDS_ID_RESPONSE_BITMASK,
  SUPPRESS_POSITIVE_RESPONSE_BIT, DEFAULT_SUBFUNCTION_MASK, and the
  DiagCommType service-prefix arrays from lib.rs into a dedicated
  cda-interfaces/src/uds.rs module. All names remain pub-re-exported
  from the crate root so no external call sites change.

- Add two transport-level helpers to uds.rs:
    - is_negative_response(&[u8]) -> bool
    - classify_pending_nrc(&[u8], source_address: u16) -> Option<UdsResponse>
  These replace the duplicated inline NRC checks in both gateways.

- Move DoipGatewaySetupError and ConnectionError from cda-interfaces
  and cda-comm-doip::lib respectively into cda-comm-doip/src/error.rs.
  cda-comm-doip re-exports both; cda-sovd now depends on cda-comm-doip
  directly for DoipGatewaySetupError.

- CAN gateway: replace the multi-arm NRC match block with a single
  classify_pending_nrc() call; use is_negative_response() to gate it.
  Handles all three NRCs (0x21, 0x78, 0x94) instead of only 0x78.

- DoIP gateway (connections.rs): replace aliased NRC imports with
  nrc:: qualified access.
