+++ cda-comm-can/Cargo.toml
@@ -48,6 +48,7 @@ tokio-socketcan-isotp = { workspace = true, optional = true }
 
 [dev-dependencies]
 toml = { workspace = true }
+cda-transport-orchestrator = { workspace = true }
 
 [features]
 default = []
