+++ cda-comm-can/src/gateway/rediscovery.rs
@@ -28,6 +28,7 @@
 
 use std::time::Duration;
 
+use cda_interfaces::EcuCanGateway;
 use tokio_util::sync::CancellationToken;
 
 use super::{CanDiagGateway, background::BackgroundTask};
