+++ cda-comm-uds/src/transport.rs
@@ -18,8 +18,8 @@ use std::{
 
 use cda_interfaces::{
     Connectivity, DiagComm, DiagServiceError, DynamicPlugin, EcuGateway, EcuManager, EcuState,
-    PayloadDecoder, ServicePayload, TransmissionParameters, UdsResponse, UdsTransport, UdsVariant,
-    VariantDetection, VariantState,
+    PayloadDecoder, PendingNrc, ServicePayload, TransmissionParameters, TransportResponse,
+    UdsTransport, UdsVariant, VariantDetection, VariantState,
     datatypes::RetryPolicy,
     diagservices::{DiagServiceResponse, UdsPayloadData},
     dlt_ctx, service_ids,
@@ -341,7 +341,7 @@ impl<S: EcuGateway, T: UdsEcuDb + VariantDetection> UdsManager<S, T> {
                 .await
                 {
                     Ok(Some(result)) => match result {
-                        Ok(Some(UdsResponse::Message(msg))) => {
+                        Ok(Some(TransportResponse::UdsResponse(msg))) => {
                             // if we received a response matching our sent SID, return it
                             // other responses are logged as warnings and ignored.
                             if !msg.data.is_empty() && msg.is_response_for_sid(sent_sid) {
@@ -361,57 +361,57 @@ impl<S: EcuGateway, T: UdsEcuDb + VariantDetection> UdsManager<S, T> {
                             }
                             tracing::warn!("Received unexpected UDS message: {:?}", msg);
                         }
-                        Ok(Some(UdsResponse::BusyRepeatRequest(_))) => {
-                            if let Err(e) = validate_timeout_by_policy(
-                                ecu_name,
-                                &uds_params.rc_21_retry_policy,
-                                &start.elapsed(),
-                                &uds_params.rc_21_completion_timeout,
-                            ) {
-                                break 'read_uds_messages Err(e);
+                        Ok(Some(TransportResponse::Pending(pending))) => match pending {
+                            PendingNrc::BusyRepeatRequest { .. } => {
+                                if let Err(e) = validate_timeout_by_policy(
+                                    ecu_name,
+                                    &uds_params.rc_21_retry_policy,
+                                    &start.elapsed(),
+                                    &uds_params.rc_21_completion_timeout,
+                                ) {
+                                    break 'read_uds_messages Err(e);
+                                }
+                                let sleep_time = uds_params.rc_21_repeat_request_time;
+                                tracing::debug!(
+                                    sleep_time = ?sleep_time,
+                                    "BusyRepeatRequest received, resending after delay"
+                                );
+                                cda_interfaces::util::tokio_ext::sleep_for(sleep_time).await;
+                                continue 'send; // continue 'send, will resend the message
                             }
-
-                            let sleep_time = uds_params.rc_21_repeat_request_time;
-                            tracing::debug!(
-                                sleep_time = ?sleep_time,
-                                "BusyRepeatRequest received, resending after delay"
-                            );
-                            cda_interfaces::util::tokio_ext::sleep_for(sleep_time).await;
-                            continue 'send; // continue 'send, will resend the message
-                        }
-                        Ok(Some(UdsResponse::TemporarilyNotAvailable(_))) => {
-                            if let Err(e) = validate_timeout_by_policy(
-                                ecu_name,
-                                &uds_params.rc_94_retry_policy,
-                                &start.elapsed(),
-                                &uds_params.rc_94_completion_timeout,
-                            ) {
-                                break 'read_uds_messages Err(e);
+                            PendingNrc::TemporarilyNotAvailable { .. } => {
+                                if let Err(e) = validate_timeout_by_policy(
+                                    ecu_name,
+                                    &uds_params.rc_94_retry_policy,
+                                    &start.elapsed(),
+                                    &uds_params.rc_94_completion_timeout,
+                                ) {
+                                    break 'read_uds_messages Err(e);
+                                }
+                                let sleep_time = uds_params.rc_94_repeat_request_time;
+                                tracing::debug!(
+                                    sleep_time = ?sleep_time,
+                                    "TemporarilyNotAvailable received, resending after delay"
+                                );
+                                cda_interfaces::util::tokio_ext::sleep_for(sleep_time).await;
+                                continue 'send; // continue 'send, will resend the message
                             }
-
-                            let sleep_time = uds_params.rc_94_repeat_request_time;
-                            tracing::debug!(
-                                sleep_time = ?sleep_time,
-                                "TemporarilyNotAvailable received, resending after delay"
-                            );
-                            cda_interfaces::util::tokio_ext::sleep_for(sleep_time).await;
-                            continue 'send; // continue 'send, will resend the message
-                        }
-                        Ok(Some(UdsResponse::ResponsePending(_))) => {
-                            if let Err(e) = validate_timeout_by_policy(
-                                ecu_name,
-                                &uds_params.rc_78_retry_policy,
-                                &start.elapsed(),
-                                &uds_params.rc_78_completion_timeout,
-                            ) {
-                                break 'read_uds_messages Err(e);
+                            PendingNrc::ResponsePending { .. } => {
+                                if let Err(e) = validate_timeout_by_policy(
+                                    ecu_name,
+                                    &uds_params.rc_78_retry_policy,
+                                    &start.elapsed(),
+                                    &uds_params.rc_78_completion_timeout,
+                                ) {
+                                    break 'read_uds_messages Err(e);
+                                }
+                                tracing::debug!(
+                                    "ResponsePending received, continue waiting for final response"
+                                );
+                                rx_timeout_next = Some(uds_params.rc_78_timeout);
+                                continue 'read_uds_messages; // continue reading UDS frames
                             }
-                            tracing::debug!(
-                                "ResponsePending received, continue waiting for final response"
-                            );
-                            rx_timeout_next = Some(uds_params.rc_78_timeout);
-                            continue 'read_uds_messages; // continue reading UDS frames
-                        }
+                        },
                         Ok(response) => {
                             break 'read_uds_messages Err(DiagServiceError::UnexpectedResponse(
                                 Some(format!("Unexpected response received: {response:?}")),
@@ -882,8 +882,8 @@ mod send_tests {
 
     use cda_interfaces::{
         DiagServiceError, EcuAddresses, EcuGateway, EcuRuntimeState, EcuStateManager, HashMap,
-        HashMapExtensions, ServicePayload, TransmissionParameters, UDS_ID_RESPONSE_BITMASK,
-        UdsResponse, VariantDetection, datatypes::FaultConfig, service_ids,
+        HashMapExtensions, PendingNrc, ServicePayload, TransmissionParameters, TransportResponse,
+        UDS_ID_RESPONSE_BITMASK, VariantDetection, datatypes::FaultConfig, service_ids,
     };
     use tokio::sync::{RwLock, mpsc};
 
@@ -931,7 +931,7 @@ mod send_tests {
     }
 
     type TestGatewaySendFn = dyn Fn(
-            mpsc::Sender<Result<Option<UdsResponse>, DiagServiceError>>,
+            mpsc::Sender<Result<Option<TransportResponse>, DiagServiceError>>,
             bool,
         ) -> Result<(), DiagServiceError>
         + Send
@@ -946,9 +946,10 @@ mod send_tests {
     /// Closures that instead want to model a gateway which closes the channel
     /// after forwarding its frame(s) (e.g. the CAN gateway breaking after the
     /// first SID-matching response) must simply let the sender drop.
-    fn park_sender(sender: mpsc::Sender<Result<Option<UdsResponse>, DiagServiceError>>) {
-        type ParkedSenders =
-            std::sync::Mutex<Vec<mpsc::Sender<Result<Option<UdsResponse>, DiagServiceError>>>>;
+    fn park_sender(sender: mpsc::Sender<Result<Option<TransportResponse>, DiagServiceError>>) {
+        type ParkedSenders = std::sync::Mutex<
+            Vec<mpsc::Sender<Result<Option<TransportResponse>, DiagServiceError>>>,
+        >;
         static PARKED: std::sync::OnceLock<ParkedSenders> = std::sync::OnceLock::new();
         PARKED
             .get_or_init(|| std::sync::Mutex::new(Vec::new()))
@@ -969,7 +970,7 @@ mod send_tests {
             &self,
             _transmission_params: TransmissionParameters,
             _message: ServicePayload,
-            response_sender: mpsc::Sender<Result<Option<UdsResponse>, DiagServiceError>>,
+            response_sender: mpsc::Sender<Result<Option<TransportResponse>, DiagServiceError>>,
             expect_uds_reply: bool,
         ) -> impl Future<Output = Result<tokio::task::JoinHandle<()>, DiagServiceError>> + Send
         {
@@ -1013,7 +1014,7 @@ mod send_tests {
             _expect_positive_response: bool,
         ) -> impl Future<
             Output = Result<
-                HashMap<String, Result<UdsResponse, DiagServiceError>>,
+                HashMap<String, Result<ServicePayload, DiagServiceError>>,
                 DiagServiceError,
             >,
         > + Send {
@@ -1050,7 +1051,7 @@ mod send_tests {
             &self,
             _transmission_params: TransmissionParameters,
             _message: ServicePayload,
-            response_sender: mpsc::Sender<Result<Option<UdsResponse>, DiagServiceError>>,
+            response_sender: mpsc::Sender<Result<Option<TransportResponse>, DiagServiceError>>,
             _expect_uds_reply: bool,
         ) -> impl Future<Output = Result<tokio::task::JoinHandle<()>, DiagServiceError>> + Send
         {
@@ -1085,7 +1086,7 @@ mod send_tests {
             _expect_positive_response: bool,
         ) -> impl Future<
             Output = Result<
-                HashMap<String, Result<UdsResponse, DiagServiceError>>,
+                HashMap<String, Result<ServicePayload, DiagServiceError>>,
                 DiagServiceError,
             >,
         > + Send {
@@ -1169,7 +1170,7 @@ mod send_tests {
     fn make_gateway() -> TestGateway {
         TestGateway {
             send_fn: Arc::new(|response_tx, _| {
-                let msg = UdsResponse::Message(ServicePayload {
+                let msg = TransportResponse::UdsResponse(ServicePayload {
                     data: vec![service_ids::SESSION_CONTROL | UDS_ID_RESPONSE_BITMASK, 0x01],
                     source_address: 0x0001,
                     target_address: 0x0E00,
@@ -1365,7 +1366,7 @@ mod send_tests {
                 // 0xF190) plus fake data. The read loop matches the SID, sees
                 // the mismatched echo bytes, ignores the frame, and loops back
                 // to recv(); the gateway task then ends and drops its sender.
-                let msg = UdsResponse::Message(ServicePayload {
+                let msg = TransportResponse::UdsResponse(ServicePayload {
                     data: vec![
                         service_ids::READ_DATA_BY_IDENTIFIER | UDS_ID_RESPONSE_BITMASK,
                         0xF2,
@@ -1458,7 +1459,7 @@ mod send_tests {
                     // First two attempts fail to transmit at all.
                     return Err(DiagServiceError::SendFailed("simulated".to_owned()));
                 }
-                let msg = UdsResponse::Message(ServicePayload {
+                let msg = TransportResponse::UdsResponse(ServicePayload {
                     data: vec![service_ids::SESSION_CONTROL | UDS_ID_RESPONSE_BITMASK, 0x01],
                     source_address: 0x0001,
                     target_address: 0x0E00,
@@ -1507,7 +1508,7 @@ mod send_tests {
         // Holds the `response_tx` handed to the *previous* attempt, so we can
         // simulate a stale gateway task pushing a late error into it after the
         // next attempt has already started.
-        type ResponseSender = mpsc::Sender<Result<Option<UdsResponse>, DiagServiceError>>;
+        type ResponseSender = mpsc::Sender<Result<Option<TransportResponse>, DiagServiceError>>;
         let prev_tx: Arc<std::sync::Mutex<Option<ResponseSender>>> =
             Arc::new(std::sync::Mutex::new(None));
         let send_count = Arc::new(std::sync::atomic::AtomicU32::new(0));
@@ -1586,7 +1587,11 @@ mod send_tests {
                 send_count_clone.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
                 // Always respond with NRC 0x21 (BusyRepeatRequest).
                 response_tx
-                    .try_send(Ok(Some(UdsResponse::BusyRepeatRequest(0x0001))))
+                    .try_send(Ok(Some(TransportResponse::Pending(
+                        PendingNrc::BusyRepeatRequest {
+                            source_address: 0x0001,
+                        },
+                    ))))
                     .ok();
                 Ok(())
             }),
@@ -1689,10 +1694,14 @@ mod send_tests {
                 let count = call_count_clone.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
                 if count == 0 {
                     response_tx
-                        .try_send(Ok(Some(UdsResponse::BusyRepeatRequest(0x0001))))
+                        .try_send(Ok(Some(TransportResponse::Pending(
+                            PendingNrc::BusyRepeatRequest {
+                                source_address: 0x0001,
+                            },
+                        ))))
                         .ok();
                 } else {
-                    let msg = UdsResponse::Message(ServicePayload {
+                    let msg = TransportResponse::UdsResponse(ServicePayload {
                         data: vec![service_ids::SESSION_CONTROL | UDS_ID_RESPONSE_BITMASK, 0x01],
                         source_address: 0x0001,
                         target_address: 0x0E00,
@@ -1731,10 +1740,14 @@ mod send_tests {
                 if count == 0 {
                     // First send ResponsePending, then the actual message
                     response_tx
-                        .try_send(Ok(Some(UdsResponse::ResponsePending(0x0001))))
+                        .try_send(Ok(Some(TransportResponse::Pending(
+                            PendingNrc::ResponsePending {
+                                source_address: 0x0001,
+                            },
+                        ))))
                         .ok();
                     response_tx
-                        .try_send(Ok(Some(UdsResponse::Message(ServicePayload {
+                        .try_send(Ok(Some(TransportResponse::UdsResponse(ServicePayload {
                             data: vec![
                                 service_ids::SESSION_CONTROL | UDS_ID_RESPONSE_BITMASK,
                                 0x01,
@@ -1774,10 +1787,14 @@ mod send_tests {
                 let count = call_count_clone.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
                 if count == 0 {
                     response_tx
-                        .try_send(Ok(Some(UdsResponse::TemporarilyNotAvailable(0x0001))))
+                        .try_send(Ok(Some(TransportResponse::Pending(
+                            PendingNrc::TemporarilyNotAvailable {
+                                source_address: 0x0001,
+                            },
+                        ))))
                         .ok();
                 } else {
-                    let msg = UdsResponse::Message(ServicePayload {
+                    let msg = TransportResponse::UdsResponse(ServicePayload {
                         data: vec![service_ids::SESSION_CONTROL | UDS_ID_RESPONSE_BITMASK, 0x01],
                         source_address: 0x0001,
                         target_address: 0x0E00,
@@ -1809,7 +1826,7 @@ mod send_tests {
         let gateway = TestGateway {
             send_fn: Arc::new(|response_tx, _| {
                 // NRC 0x7F, SID 0x10, NRC code 0x22 (conditionsNotCorrect)
-                let msg = UdsResponse::Message(ServicePayload {
+                let msg = TransportResponse::UdsResponse(ServicePayload {
                     data: vec![
                         service_ids::NEGATIVE_RESPONSE,
                         service_ids::SESSION_CONTROL,
@@ -1861,7 +1878,7 @@ mod send_tests {
     async fn test_send_with_raw_payload_sets_session_state_on_positive_response() {
         let gateway = TestGateway {
             send_fn: Arc::new(|response_tx, _| {
-                let msg = UdsResponse::Message(ServicePayload {
+                let msg = TransportResponse::UdsResponse(ServicePayload {
                     data: vec![service_ids::SESSION_CONTROL | UDS_ID_RESPONSE_BITMASK, 0x03],
                     source_address: 0x0001,
                     target_address: 0x0E00,
@@ -1939,7 +1956,7 @@ mod send_tests {
             send_fn: Arc::new(|response_tx, _| {
                 // First: a message with correct SID response but wrong DID (echo bytes)
                 // ReadDataByIdentifier (0x22) response SID is 0x62
-                let wrong_did = UdsResponse::Message(ServicePayload {
+                let wrong_did = TransportResponse::UdsResponse(ServicePayload {
                     data: vec![
                         service_ids::READ_DATA_BY_IDENTIFIER | UDS_ID_RESPONSE_BITMASK,
                         0xF2,
@@ -1953,7 +1970,7 @@ mod send_tests {
                 });
                 response_tx.try_send(Ok(Some(wrong_did))).ok();
                 // Then: the correct response with matching DID
-                let correct = UdsResponse::Message(ServicePayload {
+                let correct = TransportResponse::UdsResponse(ServicePayload {
                     data: vec![
                         service_ids::READ_DATA_BY_IDENTIFIER | UDS_ID_RESPONSE_BITMASK,
                         0xF1,
@@ -2068,7 +2085,8 @@ mod send_tests {
     async fn test_send_with_raw_payload_proceeds_after_teardown_grace_when_task_never_finishes() {
         // Far longer than RETRY_TEARDOWN_GRACE (500ms): the task handle will
         // never resolve within the scope of this test.
-        let task_delay = Duration::from_secs(600);
+        // Using 601 seconds, to triggering clippy with the smaller unit lint.
+        let task_delay = Duration::from_secs(601);
         let gateway = SlowTeardownGateway {
             task_delay,
             send_times: Arc::new(std::sync::Mutex::new(Vec::new())),
