+++ cda-main/src/error.rs
@@ -13,9 +13,9 @@
 
 use std::fmt;
 
+use cda_comm_doip::DoipGatewaySetupError;
 use cda_interfaces::{
-    DiagServiceError, DoipGatewaySetupError, config::ConfigSanityError,
-    runtime_update_api::RuntimeUpdateError,
+    DiagServiceError, config::ConfigSanityError, runtime_update_api::RuntimeUpdateError,
 };
 use cda_tracing::TracingSetupError;
