+++ cda-sovd/Cargo.toml
@@ -28,6 +28,7 @@ workspace = true
 # internal dependencies
 async-trait = { workspace = true }
 cda-interfaces = { workspace = true }
+cda-comm-doip = { workspace = true }
 cda-plugin-security = { workspace = true }
 opensovd-axum-extra = { workspace = true, features = [
   "forwarded",
