+++ cda-sovd/src/lib.rs
@@ -21,9 +21,10 @@ use axum::{
     Json,
     http::{self, Request},
 };
+use cda_comm_doip::DoipGatewaySetupError;
 use cda_interfaces::{
-    DoipGatewaySetupError, FunctionalDescriptionConfig, HashMap, SchemaProvider, UdsEcu,
-    datatypes::ComponentsConfig, dlt_ctx, file_manager::FileManager,
+    FunctionalDescriptionConfig, HashMap, SchemaProvider, UdsEcu, datatypes::ComponentsConfig,
+    dlt_ctx, file_manager::FileManager,
 };
 use cda_plugin_security::SecurityPluginLoader;
 use dynamic_router::DynamicRouter;
