+++ cda-transport-orchestrator/Cargo.toml
@@ -1,0 +1,37 @@
+# SPDX-FileCopyrightText: 2026 Copyright (c) Contributors to the Eclipse Foundation
+#
+# See the NOTICE file(s) distributed with this work for additional
+# information regarding copyright ownership.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Apache License Version 2.0 which is available at
+# https://www.apache.org/licenses/LICENSE-2.0
+#
+# SPDX-License-Identifier: Apache-2.0
+
+[package]
+name = "cda-transport-orchestrator"
+version = "0.1.0"
+edition.workspace = true
+publish = false
+description = "Transport orchestration layer for the Classic Diagnostic Adapter"
+homepage.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+# internal
+cda-interfaces = { workspace = true }
+# external
+tokio = { workspace = true, features = ["sync"] }
+tracing = { workspace = true }
+async-trait = {workspace = true}
+
+[features]
+default = []
+dlt-tracing = ["cda-interfaces/dlt-tracing"]
