+++ testcontainer/cda/Dockerfile
@@ -38,6 +38,7 @@ COPY cda-core/Cargo.toml ./cda-core/Cargo.toml
 COPY cda-comm-doip/Cargo.toml ./cda-comm-doip/Cargo.toml
 COPY cda-comm-uds/Cargo.toml ./cda-comm-uds/Cargo.toml
 COPY cda-comm-can/Cargo.toml ./cda-comm-can/Cargo.toml
+COPY cda-transport-orchestrator/Cargo.toml ./cda-transport-orchestrator/Cargo.toml
 COPY cda-database/Cargo.toml ./cda-database/Cargo.toml
 COPY cda-plugin-security/Cargo.toml ./cda-plugin-security/Cargo.toml
 COPY cda-plugin-runtime-update/Cargo.toml ./cda-plugin-runtime-update/Cargo.toml
@@ -101,6 +102,7 @@ COPY cda-core ./cda-core
 COPY cda-comm-doip ./cda-comm-doip
 COPY cda-comm-uds ./cda-comm-uds
 COPY cda-comm-can ./cda-comm-can
+COPY cda-transport-orchestrator ./cda-transport-orchestrator
 COPY cda-database ./cda-database
 COPY cda-plugin-security ./cda-plugin-security
 COPY cda-plugin-runtime-update ./cda-plugin-runtime-update
